requirements.txt
pandas==2.2.0
requests==2.29.0
streamlit==1.31.0
mplsoccer
google-generativeai==0.7.2